refactor: remove JWT and /auth as they add no value
